+++ ionoscloud/import_volume_test.go
@@ -26,7 +26,7 @@ func TestAccVolume_ImportBasic(t *testing.T) {
 				ImportStateIdFunc:       testAccVolumeImportStateId,
 				ImportState:             true,
 				ImportStateVerify:       true,
-				ImportStateVerifyIgnore: []string{"image_password", "ssh_key_path.#", "image_name"},
+				ImportStateVerifyIgnore: []string{"image_password", "ssh_key_path.#", "image_name", "user_data", "backup_unit_id"},
 			},
 		},
 	})