+++ ionoscloud/resource_group.go
@@ -182,7 +182,7 @@ func resourceGroupRead(ctx context.Context, d *schema.ResourceData, meta interfa
 	group, apiResponse, err := client.UserManagementApi.UmGroupsFindById(ctx, d.Id()).Execute()
 
 	if err != nil {
-		if apiResponse != nil && apiResponse.StatusCode == 404 {
+		if apiResponse != nil && apiResponse.Response != nil && apiResponse.StatusCode == 404 {
 			d.SetId("")
 			return nil
 		}